Update ex_stage_stubs.sv
Recreated the ALU stub package. I realized that just instantiating the packages isnt enough, as trying to drive signal through the DUT and using the ALU would do nothing, as there is no logic there. Made a new implementation that just stubs the things that are pulled from Ariane PKG ie lzc and popcount. Based entirely on the existing ALU unit with tweaks to work by itself with the Ex stage